Extract cinn from forecaster to networks. Handle soft-dependencies
